Purify Shadow on Nix for Nix flakes and reproducible & immutable builds
Required to migrate to Nix Flakes because it requires pure evalutions,
because fetchurl require a sha256 argument to be considered pure.
See https://nixos.wiki/wiki/Flakes#Making_your_evaluations_pure

A solution is to store Shadow build informations in the repository.

This also the option chosen by the Chromium NixOS maintainers:
https://github.com/NixOS/nixpkgs/tree/master/pkgs/applications/networking/browsers/chromium

However this file has to be updated regularly when new Shadow versions are released.
To avoid forgetting this, a Github Action workflow has been added to
create a Pull-request automatically in case of new versions.